pyproject.toml
[project]
name = "brickkit"
version = "0.1.0"
description = "Building blocks for Databricks Unity Catalog governance"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Casper" }]
keywords = [
    "databricks",
    "unity-catalog",
    "governance",
    "data-governance",
    "pydantic",
    "sdk",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "databricks-sdk>=0.20.0",
    "claudetracing>=0.2.0",
    "databricks-connect==17.3.*",
    "databricks-vectorsearch>=0.40",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "ty>=0.0.11",
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.urls]
# Uncomment when repo is public
# Homepage = "https://github.com/your-org/brickkit"
# Repository = "https://github.com/your-org/brickkit"
# Issues = "https://github.com/your-org/brickkit/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brickkit", "src/brickkit_tools"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff>=0.8",
    "ty>=0.0.11",
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "-ra"]
markers = [
    "unit: Unit tests (no Databricks connection)",
    "integration: Integration tests (real Databricks)",
    "slow: Long-running tests",
]
